requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
django-cors-headers==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
djoser==1.3.2
idna==2.8
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2018.7
requests==2.21.0
urllib3==1.24.1
